feat: move service tests to test folder
