+++ src/tests/services/mediaService.security.spec.ts
@@ -1,11 +1,11 @@
 import { PrismaClient } from "@prisma/client";
-import createUserAction from "../actions/createUserAction";
-import createProjectAction from "../actions/createProjectAction";
-import createStatusAction from "../actions/createStatusAction";
-import ticketService from "./ticketsService";
-import createTicketAction from "../actions/createTicketAction";
-import createMediaAction from "../actions/createMediaAction";
-import mediaService from "./mediaService";
+import createUserAction from "../../actions/createUserAction";
+import createProjectAction from "../../actions/createProjectAction";
+import createStatusAction from "../../actions/createStatusAction";
+import createMediaAction from "../../actions/createMediaAction";
+import mediaService from "../../services/mediaService";
+import createTicketAction from "../../actions/createTicketAction";
+import projectService from "../../services/projectsService";
 const faker = require("@faker-js/faker");
 
 const prisma = new PrismaClient();
@@ -14,10 +14,18 @@ afterAll(() => {
   return expect(prisma.$disconnect()).resolves;
 });
 
-describe("mediaService", () => {
-  it("creates new media correctly", async () => {
+describe("mediaService - security", () => {
+  it("throw an error if non-member user trying to get a media", async () => {
     // create a user, role, project, status, ticket and media
-    const savedUser = await createUserAction({
+    const savedUser1 = await createUserAction({
+      profilePicture: faker.image.people(500, 500),
+      firstName: faker.name.firstName(),
+      lastName: faker.name.lastName(),
+      email: faker.internet.email(),
+      password: faker.internet.password(),
+    });
+
+    const savedUser2 = await createUserAction({
       profilePicture: faker.image.people(500, 500),
       firstName: faker.name.firstName(),
       lastName: faker.name.lastName(),
@@ -30,62 +38,7 @@ describe("mediaService", () => {
       description: faker.random.words(5),
       createdAt: faker.date.recent(),
       estimateEndAt: faker.date.future(),
-      userId: savedUser.id,
-    });
-
-    const savedStatus = await createStatusAction({
-      name: faker.random.word(),
-    });
-
-    const name = `${faker.hacker.verb()} ${faker.hacker.adjective()} ${faker.hacker.noun()}`;
-    const description = faker.random.words(5);
-    const createdAt = faker.date.recent();
-
-    const savedTicket = await ticketService.create({
-      name,
-      description,
-      projectId: savedProject.id,
-      statusId: savedStatus.id,
-      assigneeId: savedUser.id,
-      createdAt,
-    });
-
-    const mediaName = "test.jpg";
-    const mediaType = "image/jpeg";
-    const mediaUrl = faker.image.imageUrl();
-
-    const savedMedia = await mediaService.create(
-      {
-        name: mediaName,
-        type: mediaType,
-        url: mediaUrl,
-        createdAt: createdAt,
-        ticketId: savedTicket.id,
-      },
-      savedUser.id
-    );
-
-    expect(savedMedia.name).toBe(mediaName);
-    expect(savedMedia.type).toBe(mediaType);
-    expect(savedMedia.url).toBe(mediaUrl);
-    expect(savedMedia.ticketId).toBe(savedTicket.id);
-  });
-
-  it("delete a media correctly", async () => {
-    const savedUser = await createUserAction({
-      profilePicture: faker.image.people(500, 500),
-      firstName: faker.name.firstName(),
-      lastName: faker.name.lastName(),
-      email: faker.internet.email(),
-      password: faker.internet.password(),
-    });
-
-    const savedProject = await createProjectAction({
-      name: faker.internet.domainName(),
-      description: faker.random.words(5),
-      createdAt: faker.date.recent(),
-      estimateEndAt: faker.date.future(),
-      userId: savedUser.id,
+      userId: savedUser1.id,
     });
 
     const savedStatus = await createStatusAction({
@@ -101,7 +54,7 @@ describe("mediaService", () => {
       description,
       projectId: savedProject.id,
       statusId: savedStatus.id,
-      assigneeId: savedUser.id,
+      assigneeId: savedUser1.id,
       createdAt,
     });
 
@@ -109,81 +62,155 @@ describe("mediaService", () => {
     const mediaType = "image/jpeg";
     const mediaUrl = faker.image.imageUrl();
 
-    const savedMedia = await mediaService.create(
-      {
-        name: mediaName,
-        type: mediaType,
-        url: mediaUrl,
-        createdAt: createdAt,
-        ticketId: savedTicket.id,
-      },
-      savedUser.id
-    );
-
-    await mediaService.delete(savedMedia.id, savedUser.id);
-
-    const deletedMedia = await prisma.media.findUnique({
-      where: { id: savedMedia.id },
-    });
-
-    expect(deletedMedia).toBeNull();
-  });
-
-  it("get ticket media correctly", async () => {
-    const savedUser = await createUserAction({
-      profilePicture: faker.image.people(500, 500),
-      firstName: faker.name.firstName(),
-      lastName: faker.name.lastName(),
-      email: faker.internet.email(),
-      password: faker.internet.password(),
-    });
-
-    const savedProject = await createProjectAction({
-      name: faker.internet.domainName(),
-      description: faker.random.words(5),
-      createdAt: faker.date.recent(),
-      estimateEndAt: faker.date.future(),
-      userId: savedUser.id,
-    });
-
-    const savedStatus = await createStatusAction({
-      name: faker.random.word(),
-    });
-
-    const name = `${faker.hacker.verb()} ${faker.hacker.adjective()} ${faker.hacker.noun()}`;
-    const description = faker.random.words(5);
-    const createdAt = faker.date.recent();
-
-    const savedTicket = await createTicketAction({
-      name,
-      description,
-      projectId: savedProject.id,
-      statusId: savedStatus.id,
-      assigneeId: savedUser.id,
-      createdAt,
-    });
-
     await createMediaAction({
-      name: "test.jpg",
-      type: "image/jpeg",
-      url: faker.image.imageUrl(),
-      createdAt: createdAt,
-      ticketId: savedTicket.id,
-    });
-    await createMediaAction({
-      name: "test2.jpg",
-      type: "image/jpeg",
-      url: faker.image.imageUrl(),
+      name: mediaName,
+      type: mediaType,
+      url: mediaUrl,
       createdAt: createdAt,
       ticketId: savedTicket.id,
     });
 
-    const ticketMedia = await mediaService.getTicketMedia(
+    const mediaPromise = mediaService.getTicketMedia(
       savedTicket.id,
-      savedUser.id
+      savedUser2.id
     );
 
-    expect(ticketMedia).toBeTruthy();
-    expect(ticketMedia).toHaveLength(2);
+    await expect(mediaPromise).rejects.toThrow(
+      "User is not a member of the project"
+    );
+  });
+
+  it("throw an error if non-member user trying to create a media", async () => {
+    // create a user, role, project, status, ticket and media
+    const savedUser1 = await createUserAction({
+      profilePicture: faker.image.people(500, 500),
+      firstName: faker.name.firstName(),
+      lastName: faker.name.lastName(),
+      email: faker.internet.email(),
+      password: faker.internet.password(),
+    });
+
+    const savedUser2 = await createUserAction({
+      profilePicture: faker.image.people(500, 500),
+      firstName: faker.name.firstName(),
+      lastName: faker.name.lastName(),
+      email: faker.internet.email(),
+      password: faker.internet.password(),
+    });
+
+    const savedProject = await createProjectAction({
+      name: faker.internet.domainName(),
+      description: faker.random.words(5),
+      createdAt: faker.date.recent(),
+      estimateEndAt: faker.date.future(),
+      userId: savedUser1.id,
+    });
+
+    const savedStatus = await createStatusAction({
+      name: faker.random.word(),
+    });
+
+    const name = `${faker.hacker.verb()} ${faker.hacker.adjective()} ${faker.hacker.noun()}`;
+    const description = faker.random.words(5);
+    const createdAt = faker.date.recent();
+
+    const savedTicket = await createTicketAction({
+      name,
+      description,
+      projectId: savedProject.id,
+      statusId: savedStatus.id,
+      assigneeId: savedUser1.id,
+      createdAt,
+    });
+
+    const mediaName = "test.jpg";
+    const mediaType = "image/jpeg";
+    const mediaUrl = faker.image.imageUrl();
+
+    const mediaPromise = mediaService.create(
+      {
+        name: mediaName,
+        type: mediaType,
+        url: mediaUrl,
+        createdAt: createdAt,
+        ticketId: savedTicket.id,
+      },
+      savedUser2.id
+    );
+
+    await expect(mediaPromise).rejects.toThrow(
+      "User is not a member of the project"
+    );
+  });
+
+  it("throw an error if user trying to delete a media is not project admin", async () => {
+    // create a user, role, project, status, ticket and media
+    const savedUser1 = await createUserAction({
+      profilePicture: faker.image.people(500, 500),
+      firstName: faker.name.firstName(),
+      lastName: faker.name.lastName(),
+      email: faker.internet.email(),
+      password: faker.internet.password(),
+    });
+
+    const savedUser2 = await createUserAction({
+      profilePicture: faker.image.people(500, 500),
+      firstName: faker.name.firstName(),
+      lastName: faker.name.lastName(),
+      email: faker.internet.email(),
+      password: faker.internet.password(),
+    });
+
+    const savedProject = await createProjectAction({
+      name: faker.internet.domainName(),
+      description: faker.random.words(5),
+      createdAt: faker.date.recent(),
+      estimateEndAt: faker.date.future(),
+      userId: savedUser1.id,
+    });
+
+    await projectService.addProjectMember({
+      roleId: 2,
+      userId: savedUser2.id,
+      projectId: savedProject.id,
+    });
+
+    const savedStatus = await createStatusAction({
+      name: faker.random.word(),
+    });
+
+    const name = `${faker.hacker.verb()} ${faker.hacker.adjective()} ${faker.hacker.noun()}`;
+    const description = faker.random.words(5);
+    const createdAt = faker.date.recent();
+
+    const savedTicket = await createTicketAction({
+      name,
+      description,
+      projectId: savedProject.id,
+      statusId: savedStatus.id,
+      assigneeId: savedUser1.id,
+      createdAt,
+    });
+
+    const mediaName = "test.jpg";
+    const mediaType = "image/jpeg";
+    const mediaUrl = faker.image.imageUrl();
+
+    const media = await mediaService.create(
+      {
+        name: mediaName,
+        type: mediaType,
+        url: mediaUrl,
+        createdAt: createdAt,
+        ticketId: savedTicket.id,
+      },
+      savedUser1.id
+    );
+
+    const mediaPromise = mediaService.delete(media.id, savedUser2.id);
+
+    await expect(mediaPromise).rejects.toThrow(
+      "User is not the project Administrator"
+    );
   });
 });
